% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten.R
\name{guten_book}
\alias{guten_book}
\title{Get details for a specific Gutenberg book}
\usage{
guten_book(id)
}
\arguments{
\item{id}{Gutenberg book ID (integer)}
}
\value{
tibble with one row: id, title, authors, languages, download_count, subjects, bookshelves
}
\description{
Get details for a specific Gutenberg book
}
