% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten.R
\name{guten_context}
\alias{guten_context}
\title{Show Gutendex package context for LLM use}
\usage{
guten_context()
}
\value{
Invisible string with full context
}
\description{
Show Gutendex package context for LLM use
}
