% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten.R
\name{guten_search}
\alias{guten_search}
\title{Search the Gutendex book catalog}
\usage{
guten_search(query, page = 1)
}
\arguments{
\item{query}{Search term (title, author, etc.)}

\item{page}{Page number (default 1, 32 results per page)}
}
\value{
tibble: id, title, authors, languages, download_count, subjects, bookshelves
}
\description{
Search the Gutendex book catalog
}
